backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
pypdf==5.0.1
anthropic>=0.49.0
pydantic==2.9.2
python-dotenv==1.0.1
